refactor: extract fragmentToDoc to its own file
Lets unit tests import the pure parser without pulling in React via
canvasStorage.ts. Re-exported from canvasStorage.ts so existing
consumers are unchanged.
