+++ client/src/components/pages/developer/CalculatorCanvas/canvasStorage.ts
@@ -48,66 +48,8 @@ export {
 
 // ─── Serialise / deserialise ──────────────────────────────────────────────────
 
-export function fragmentToDoc(f: RateBuildupTemplateFullSnippetFragment): CanvasDocument {
-  const specialPositions: SpecialNodePositions = {
-    quantity: { x: 100, y: 200 },
-    unitPrice: { x: 700, y: 200 },
-  };
-  if (f.specialPositions) {
-    try {
-      const sp = typeof f.specialPositions === "string"
-        ? JSON.parse(f.specialPositions)
-        : f.specialPositions;
-      if (sp?.quantity) specialPositions.quantity = sp.quantity;
-      if (sp?.unitPrice) specialPositions.unitPrice = sp.unitPrice;
-    } catch { /* ignore */ }
-  }
-
-  // groupDefs and controllerDefs arrive as typed sub-docs after server migration;
-  // fall back to parsing JSON strings for pre-migration data.
-  let groupDefs: GroupDef[] = [];
-  if (Array.isArray(f.groupDefs)) {
-    groupDefs = f.groupDefs as GroupDef[];
-  } else if (typeof f.groupDefs === "string") {
-    try { groupDefs = JSON.parse(f.groupDefs); } catch { /* ignore */ }
-  }
-  // Deduplicate memberIds within each group
-  groupDefs = groupDefs.map((g) => ({ ...g, memberIds: [...new Set(g.memberIds)] }));
-
-  let controllerDefs: ControllerDef[] = [];
-  if (Array.isArray(f.controllerDefs)) {
-    controllerDefs = f.controllerDefs as ControllerDef[];
-  } else if (typeof f.controllerDefs === "string") {
-    try { controllerDefs = JSON.parse(f.controllerDefs); } catch { /* ignore */ }
-  }
-
-  // Deduplicate defs by ID — protects against duplicate entries that may have been
-  // saved when the canvas had a bug where content edits didn't visually reflect.
-  const dedup = <T extends { id: string }>(arr: T[]): T[] => {
-    const seen = new Set<string>();
-    return arr.filter((item) => {
-      if (seen.has(item.id)) return false;
-      seen.add(item.id);
-      return true;
-    });
-  };
-
-  return {
-    id: f._id,
-    label: f.label,
-    defaultUnit: f.defaultUnit ?? "unit",
-    parameterDefs: dedup((f.parameterDefs ?? []) as CanvasParameterDef[]),
-    tableDefs: dedup((f.tableDefs ?? []) as CanvasTableDef[]),
-    formulaSteps: dedup((f.formulaSteps ?? []) as CanvasFormulaStep[]),
-    breakdownDefs: dedup((f.breakdownDefs ?? []) as CanvasBreakdownDef[]),
-    outputDefs: dedup(((f.outputDefs ?? []) as OutputDef[])),
-    specialPositions,
-    groupDefs,
-    controllerDefs,
-    unitVariants: (f.unitVariants ?? []).map(({ unit, activatesGroupId, conversionFormula }) => ({ unit, activatesGroupId, conversionFormula: conversionFormula ?? undefined })),
-    updatedAt: f.updatedAt ?? undefined,
-  };
-}
+export { fragmentToDoc } from "./fragmentParser";
+import { fragmentToDoc } from "./fragmentParser";
 
 function omitTypename<T>(obj: T): T {
   if (Array.isArray(obj)) return obj.map(omitTypename) as unknown as T;
